Misc cleanups from kernel refactor
