+++ packages/dds/tree/src/test/codec/ajvValidator.ts
@@ -3,10 +3,6 @@
  * Licensed under the MIT License.
  */
 
-import type { IRequest } from "@fluidframework/core-interfaces";
-import type { IFluidHandleContext } from "@fluidframework/core-interfaces/internal";
-import { create404Response } from "@fluidframework/runtime-utils/internal";
-import { FluidSerializer } from "@fluidframework/shared-object-base/internal";
 import { MockHandle } from "@fluidframework/test-runtime-utils/internal";
 import type { Static, TSchema } from "@sinclair/typebox";
 // Based on ESM workaround from https://github.com/ajv-validator/ajv/issues/2047#issuecomment-1241470041 .
@@ -22,6 +18,7 @@ const Ajv =
 	(ajvModuleOrClass as any);
 
 import type { JsonValidator } from "../../codec/index.js";
+import { mockSerializer } from "../mockSerializer.js";
 
 // See: https://github.com/sinclairzx81/typebox#ajv
 const ajv = formats.default(new Ajv({ strict: false, allErrors: true }), [
@@ -41,28 +38,6 @@ const ajv = formats.default(new Ajv({ strict: false, allErrors: true }), [
 	"regex",
 ]);
 
-class MockHandleContext implements IFluidHandleContext {
-	public isAttached = false;
-	public get IFluidHandleContext() {
-		return this;
-	}
-
-	public constructor(
-		public readonly absolutePath = "",
-		public readonly routeContext?: IFluidHandleContext,
-	) {}
-
-	public attachGraph() {
-		throw new Error("Method not implemented.");
-	}
-
-	public async resolveHandle(request: IRequest) {
-		return create404Response(request);
-	}
-}
-
-const serializer = new FluidSerializer(new MockHandleContext());
-
 /**
  * A {@link JsonValidator} implementation which uses Ajv's JSON schema validator.
  *
@@ -77,7 +52,7 @@ export const ajvValidator: JsonValidator = {
 				const valid = validate(data);
 				if (!valid) {
 					throw new Error(
-						`Invalid JSON.\n\nData: ${serializer.stringify(
+						`Invalid JSON.\n\nData: ${mockSerializer.stringify(
 							data,
 							new MockHandle(""),
 						)}\n\nErrors: ${JSON.stringify(validate.errors)}`,
